move validation fns to separate plugin py
